notification-service/pyproject.toml
[tool.poetry]
name = "notification-service"
version = "0.1.0"
description = "Notifications API"
authors = ["Kostiantyn Ivashchenko <kivaschenko@protonmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
uvicorn = "^0.30.6"
pydantic = "^2.8.2"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
